chore: rename stories
