+++ apps/compositions/src/examples/code-block-basic.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { CodeBlock, type CodeBlockAdapter, IconButton } from "@chakra-ui/react"
+import { CodeBlock, type CodeBlockAdapter } from "@chakra-ui/react"
 import type { HighlighterGeneric } from "shiki"
 
 const file = {
@@ -13,18 +13,10 @@ const file = {
   title: "index.html",
 }
 
-export const CodeBlockShikiWithTabsSync = () => {
+export const CodeBlockBasic = () => {
   return (
     <CodeBlock.AdapterProvider value={shikiAdapter}>
       <CodeBlock.Root code={file.code} language={file.language}>
-        <CodeBlock.Header>
-          <CodeBlock.Title>{file.title}</CodeBlock.Title>
-          <CodeBlock.CopyTrigger asChild>
-            <IconButton variant="ghost" size="2xs">
-              <CodeBlock.CopyIndicator />
-            </IconButton>
-          </CodeBlock.CopyTrigger>
-        </CodeBlock.Header>
         <CodeBlock.Content>
           <CodeBlock.Code>
             <CodeBlock.CodeText />
@@ -40,7 +32,7 @@ const shikiAdapter: CodeBlockAdapter = {
     const { createHighlighter } = await import("shiki")
     return createHighlighter({
       langs: ["tsx", "scss", "html", "bash", "json"],
-      themes: ["github-dark"],
+      themes: ["github-dark", "github-light"],
     })
   },
   getHighlighter: (ctx: HighlighterGeneric<any, any> | null) => {
@@ -54,7 +46,8 @@ const shikiAdapter: CodeBlockAdapter = {
         code: removeWrapperTags(
           ctx.codeToHtml(code, {
             lang: language,
-            theme: "github-dark",
+            theme:
+              meta?.colorScheme === "dark" ? "github-dark" : "github-light",
             transformers: [
               {
                 line(hast, line) {
@@ -65,6 +58,14 @@ const shikiAdapter: CodeBlockAdapter = {
                       ? ""
                       : undefined,
                     "data-word-wrap": meta?.wordWrap ? "" : undefined,
+                    "data-diff": meta?.addedLineNumbers?.includes(line)
+                      ? "added"
+                      : meta?.removedLineNumbers?.includes(line)
+                        ? "removed"
+                        : undefined,
+                    "data-focused": meta?.focusedLineNumbers?.includes(line)
+                      ? ""
+                      : undefined,
                   })
                 },
               },
