+++ apps/compositions/src/examples/code-block-with-custom-action.tsx
@@ -1,6 +1,7 @@
 "use client"
 
 import { CodeBlock, type CodeBlockAdapter, IconButton } from "@chakra-ui/react"
+import { RxCodesandboxLogo } from "react-icons/rx"
 import type { HighlighterGeneric } from "shiki"
 
 const file = {
@@ -13,7 +14,7 @@ const file = {
   title: "index.html",
 }
 
-export const CodeBlockShikiWithCopyButton = () => {
+export const CodeBlockWithCustomAction = () => {
   return (
     <CodeBlock.AdapterProvider value={shikiAdapter}>
       <CodeBlock.Root code={file.code} language={file.language}>
@@ -24,6 +25,13 @@ export const CodeBlockShikiWithCopyButton = () => {
               <CodeBlock.CopyIndicator />
             </IconButton>
           </CodeBlock.CopyTrigger>
+          <IconButton
+            variant="ghost"
+            size="2xs"
+            aria-label="Open in CodeSandbox"
+          >
+            <RxCodesandboxLogo />
+          </IconButton>
         </CodeBlock.Header>
         <CodeBlock.Content>
           <CodeBlock.Code>
