+++ apps/compositions/src/examples/code-block-with-sizes.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { CodeBlock, type CodeBlockAdapter, IconButton } from "@chakra-ui/react"
+import { CodeBlock, type CodeBlockAdapter, For, Stack } from "@chakra-ui/react"
 import type { HighlighterGeneric } from "shiki"
 
 const file = {
@@ -13,24 +13,30 @@ const file = {
   title: "index.html",
 }
 
-export const CodeBlockShikiWithTabs = () => {
+export const CodeBlockWithSizes = () => {
   return (
     <CodeBlock.AdapterProvider value={shikiAdapter}>
-      <CodeBlock.Root code={file.code} language={file.language}>
-        <CodeBlock.Header>
-          <CodeBlock.Title>{file.title}</CodeBlock.Title>
-          <CodeBlock.CopyTrigger asChild>
-            <IconButton variant="ghost" size="2xs">
-              <CodeBlock.CopyIndicator />
-            </IconButton>
-          </CodeBlock.CopyTrigger>
-        </CodeBlock.Header>
-        <CodeBlock.Content>
-          <CodeBlock.Code>
-            <CodeBlock.CodeText />
-          </CodeBlock.Code>
-        </CodeBlock.Content>
-      </CodeBlock.Root>
+      <Stack gap="8">
+        <For each={["sm", "md", "lg"]}>
+          {(size) => (
+            <CodeBlock.Root
+              key={size}
+              code={file.code}
+              language={file.language}
+              size={size}
+            >
+              <CodeBlock.Header>
+                <CodeBlock.Title>(size={size})</CodeBlock.Title>
+              </CodeBlock.Header>
+              <CodeBlock.Content>
+                <CodeBlock.Code>
+                  <CodeBlock.CodeText />
+                </CodeBlock.Code>
+              </CodeBlock.Content>
+            </CodeBlock.Root>
+          )}
+        </For>
+      </Stack>
     </CodeBlock.AdapterProvider>
   )
 }
@@ -40,7 +46,7 @@ const shikiAdapter: CodeBlockAdapter = {
     const { createHighlighter } = await import("shiki")
     return createHighlighter({
       langs: ["tsx", "scss", "html", "bash", "json"],
-      themes: ["github-dark"],
+      themes: ["github-dark", "github-light"],
     })
   },
   getHighlighter: (ctx: HighlighterGeneric<any, any> | null) => {
@@ -54,7 +60,8 @@ const shikiAdapter: CodeBlockAdapter = {
         code: removeWrapperTags(
           ctx.codeToHtml(code, {
             lang: language,
-            theme: "github-dark",
+            theme:
+              meta?.colorScheme === "dark" ? "github-dark" : "github-light",
             transformers: [
               {
                 line(hast, line) {
@@ -65,6 +72,14 @@ const shikiAdapter: CodeBlockAdapter = {
                       ? ""
                       : undefined,
                     "data-word-wrap": meta?.wordWrap ? "" : undefined,
+                    "data-diff": meta?.addedLineNumbers?.includes(line)
+                      ? "added"
+                      : meta?.removedLineNumbers?.includes(line)
+                        ? "removed"
+                        : undefined,
+                    "data-focused": meta?.focusedLineNumbers?.includes(line)
+                      ? ""
+                      : undefined,
                   })
                 },
               },
