+++ apps/compositions/src/examples/code-block-with-tabs-sync.tsx
@@ -1,7 +1,6 @@
 "use client"
 
 import { CodeBlock, type CodeBlockAdapter, IconButton } from "@chakra-ui/react"
-import { RxCodesandboxLogo } from "react-icons/rx"
 import type { HighlighterGeneric } from "shiki"
 
 const file = {
@@ -14,7 +13,7 @@ const file = {
   title: "index.html",
 }
 
-export const CodeBlockShikiWithCustomAction = () => {
+export const CodeBlockWithTabsSync = () => {
   return (
     <CodeBlock.AdapterProvider value={shikiAdapter}>
       <CodeBlock.Root code={file.code} language={file.language}>
@@ -25,13 +24,6 @@ export const CodeBlockShikiWithCustomAction = () => {
               <CodeBlock.CopyIndicator />
             </IconButton>
           </CodeBlock.CopyTrigger>
-          <IconButton
-            variant="ghost"
-            size="2xs"
-            aria-label="Open in CodeSandbox"
-          >
-            <RxCodesandboxLogo />
-          </IconButton>
         </CodeBlock.Header>
         <CodeBlock.Content>
           <CodeBlock.Code>
@@ -73,11 +65,6 @@ const shikiAdapter: CodeBlockAdapter = {
                       ? ""
                       : undefined,
                     "data-word-wrap": meta?.wordWrap ? "" : undefined,
-                    "data-diff": meta?.addedLineNumbers?.includes(line)
-                      ? "added"
-                      : meta?.removedLineNumbers?.includes(line)
-                        ? "removed"
-                        : undefined,
                   })
                 },
               },
