+++ packages/react/__stories__/code-block.stories.tsx
@@ -12,17 +12,18 @@ export default {
   ],
 } satisfies Meta
 
-export { CodeBlockShiki as Shiki } from "compositions/examples/code-block-shiki"
-export { CodeBlockShikiWithCopyButton as CopyButton } from "compositions/examples/code-block-shiki-with-copy-button"
-export { CodeBlockShikiWithCustomAction as CustomAction } from "compositions/examples/code-block-shiki-with-custom-action"
-export { CodeBlockShikiWithLineFocusing as LineFocusing } from "compositions/examples/code-block-shiki-with-line-focusing"
-export { CodeBlockShikiWithLineHighlighting as LineHighlighting } from "compositions/examples/code-block-shiki-with-line-highlighting"
-export { CodeBlockShikiWithTitle as Title } from "compositions/examples/code-block-shiki-with-title"
-export { CodeBlockShikiWithLineNumbers as LineNumbers } from "compositions/examples/code-block-shiki-with-line-numbers"
-export { CodeBlockShikiWithDiff as Diff } from "compositions/examples/code-block-shiki-with-diff"
-export { CodeBlockShikiWithMaxLines as MaxLines } from "compositions/examples/code-block-shiki-with-max-lines"
-export { CodeBlockHighlightJs as HighlightJs } from "compositions/examples/code-block-highlight-js"
-export { CodeBlockShikiWithWordWrap as WordWrap } from "compositions/examples/code-block-shiki-with-word-wrap"
-export { CodeBlockShikiWithLanguageSwitcher as LanguageSwitcher } from "compositions/examples/code-block-shiki-with-language-switcher"
-export { CodeBlockShikiWithThemes as Themes } from "compositions/examples/code-block-shiki-with-themes"
-export { CodeBlockShikiWithSizes as Sizes } from "compositions/examples/code-block-shiki-with-sizes"
+export { CodeBlockBasic as Basic } from "compositions/examples/code-block-basic"
+export { CodeBlockWithCopyButton as CopyButton } from "compositions/examples/code-block-with-copy-button"
+export { CodeBlockWithCustomAction as CustomAction } from "compositions/examples/code-block-with-custom-action"
+export { CodeBlockWithLineFocus as LineFocusing } from "compositions/examples/code-block-with-line-focus"
+export { CodeBlockWithLineHighlight as LineHighlighting } from "compositions/examples/code-block-with-line-highlight"
+export { CodeBlockWithTitle as Title } from "compositions/examples/code-block-with-title"
+export { CodeBlockWithLineNumbers as LineNumbers } from "compositions/examples/code-block-with-line-numbers"
+export { CodeBlockWithDiff as Diff } from "compositions/examples/code-block-with-diff"
+export { CodeBlockWithMaxLines as MaxLines } from "compositions/examples/code-block-with-max-lines"
+export { CodeBlockWithWordWrap as WordWrap } from "compositions/examples/code-block-with-word-wrap"
+export { CodeBlockWithLanguageSwitcher as LanguageSwitcher } from "compositions/examples/code-block-with-language-switcher"
+export { CodeBlockWithSizes as Sizes } from "compositions/examples/code-block-with-sizes"
+export { CodeBlockWithFloatingAction as FloatingAction } from "compositions/examples/code-block-with-floating-action"
+export { CodeBlockWithThemes as Themes } from "compositions/examples/code-block-with-themes"
+export { CodeBlockWithHighlightJs as HighlightJs } from "compositions/examples/code-block-with-highlight-js"
